requirements.txt
aenum
bidict>=0.18.0
munch
click
python-rapidjson
fs
deepdiff
rtree
pydantic>=1.8.0
shapely>=1.8.0
geopandas>=0.8.1
numpy
scikit-image
scikit-learn==0.23.2
joblib
lightgbm>=3.0.0
sentinelhub>=3.4.0
eo-learn-core>=0.10.0
eo-learn-io>=0.10.0
eo-learn-features>=0.10.0
eo-learn-mask>=0.10.0
eo-learn-ml-tools>=0.10.0
eo-learn-geometry>=0.10.0
eo-learn-visualization>=0.10.0
ray[default]
